+++ android/AdMob/app/src/main/java/com/pozirk/ads/admob/context/ExtensionContext.java
@@ -14,10 +14,19 @@
  * limitations under the License.
  */
 
-package com.pozirk.ads.admob;
+package com.pozirk.ads.admob.context;
 
 import com.adobe.fre.FREContext;
 import com.adobe.fre.FREFunction;
+import com.pozirk.ads.admob.manager.AdMobManager;
+import com.pozirk.ads.admob.function.CacheInterstitialFunction;
+import com.pozirk.ads.admob.function.DisposeFunction;
+import com.pozirk.ads.admob.function.HideFunction;
+import com.pozirk.ads.admob.function.InitFunction;
+import com.pozirk.ads.admob.function.SetVolumeFunction;
+import com.pozirk.ads.admob.function.ShowFunction;
+import com.pozirk.ads.admob.function.ShowInterstitialFunction;
+
 import java.util.HashMap;
 import java.util.Map;
