+++ android/AdMob/app/src/main/java/com/pozirk/ads/admob/function/CacheInterstitialFunction.java
@@ -14,20 +14,31 @@
  * limitations under the License.
  */
 
-package com.pozirk.ads.admob;
+package com.pozirk.ads.admob.function;
 
 import com.adobe.fre.FREContext;
 import com.adobe.fre.FREFunction;
 import com.adobe.fre.FREObject;
+import com.pozirk.ads.admob.context.ExtensionContext;
 
-public class ShowInterstitialFunction
+public class CacheInterstitialFunction
   implements FREFunction
-{	
+{
 	public FREObject call(FREContext frectx, FREObject[] args)
   {
   	ExtensionContext ctx = (ExtensionContext)frectx;
 
-  	ctx.getAdMobMan().showInterstitial();
+  	try
+    {
+  		FREObject adID = args[0];
+    	FREObject testDevice = args[1];
+    	ctx.getAdMobMan().cacheInterstitial(adID.getAsString(), (testDevice != null ? testDevice.getAsString() : null));
+    }
+  	catch(Exception e)
+    {
+    	e.printStackTrace();
+    	ctx.dispatchStatusEventAsync("INTERSTITIAL_CACHE_FAIL", e.getMessage());
+    }
 
   	return null;
   }
